Stop attempting to zone on zone request. Don't attempt to follow master out of mog house.
Handles zoning into a mog house over a zone line.
Mog houses are not in a different zone. Instead of waiting for zone change
to consider zoning complete, instead use the client's 'zone request' packet
as the trigger to stop trying to zone. Probably the clean way to do this.

Handles zoning out of a mog house. These trigger the 'zone request' on the master,
but mules cannot zone out via just running at a zone line. For now, mules will
not follow thier master out of the mog house.