Mitigate against a race condition whilst zoning.
The race condition is receiving a zone request from the character being
followed, after intercepting the outgoing packet from the client to zone.
This is eliminated in most cases by ceasing zone attempts when the zone
changes. In some cases (e.g. into mog houses), the zone does not change.